% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CosmxHumanLungP9S1.R
\name{CosmxLungP9S1}
\alias{CosmxLungP9S1}
\title{Obtain the CosMx data}
\usage{
CosmxLungP9S1()
}
\value{
A \linkS4class{SpatialExperiment} object with a matrix of UMI counts
and spatial coordinates
}
\description{
Obtain the original CosMx lung patient 9 slice 1 data from:
https://nanostring.com/resources/smi-ffpe-dataset-lung9-rep1-data/
}
\details{
\itemize{
\item exprMat_file.csv
\item metadata_file.csv
}
}
\references{
TODO: how to deal with not publication?
}
\author{
Estella Yixing Dong
}
