% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MerscopeHumanOvarianP2S2.R
\name{MerscopeHumanOvarianP2S2}
\alias{MerscopeHumanOvarianP2S2}
\title{Obtain the MERSCOPE data}
\usage{
MerscopeHumanOvarianP2S2()
}
\value{
A \linkS4class{SpatialExperiment} object with a matrix of UMI counts
and spatial coordinates
}
\description{
Obtain the original MERSCOPE ovarian patient 2 slice 2 data from:
https://info.vizgen.com/ffpe-showcase
}
\details{
\itemize{
\item cell_by_gene.csv
https://console.cloud.google.com/storage/browser/_details/vz-ffpe-showcase/HumanOvarianCancerPatient2Slice2/cell_by_gene.csv?pageState=(\%22StorageObjectListTable\%22:(\%22f\%22:\%22\%255B\%255D\%22))
\item cell_metadata.csv
https://console.cloud.google.com/storage/browser/_details/vz-ffpe-showcase/HumanOvarianCancerPatient2Slice2/cell_metadata.csv?pageState=(\%22StorageObjectListTable\%22:(\%22f\%22:\%22\%255B\%255D\%22))
}
}
\references{
TODO: how to deal with not publication?
}
\author{
Estella Yixing Dong
}
