% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XeniumMouseBrainSubset.R
\name{XeniumMouseBrainSubset}
\alias{XeniumMouseBrainSubset}
\title{Obtain the Xenium data}
\usage{
XeniumMouseBrainSubset()
}
\value{
A \linkS4class{SpatialExperiment} object with a matrix of UMI counts
and spatial coordinates
}
\description{
Obtain the original Xenium Mouse Brain tiny subset data from:
https://cf.10xgenomics.com/samples/xenium/1.0.2/Xenium_V1_FF_Mouse_Brain_Coronal_Subset_CTX_HP/Xenium_V1_FF_Mouse_Brain_Coronal_Subset_CTX_HP_outs.zip
}
\details{
\itemize{
\item exprMat_file.csv
\item metadata_file.csv
}
}
\references{
TODO: how to deal with not publication?
}
\author{
Estella Yixing Dong
}
